+++ server/backend/controllers/productTypeController.js
@@ -80,9 +80,11 @@ const updateProductType = requestHandler(async (req, res, database) => {
     let newImage = currentImage;
     if(image) {
         newImage = image;
-        const filePath = `uploads/producttypes/${currentImage}`;
-        const isFileExists = await fileExists(filePath);
-        if(isFileExists) await unlink(getDir(filePath));
+        if(currentImage) {
+            const filePath = `uploads/producttypes/${currentImage}`;
+            const isFileExists = await fileExists(filePath);
+            if(isFileExists) await unlink(getDir(filePath));
+        }
     }
     const [update] = await database.execute(productTypeStmt.updateProductType, [name, newImage, id]);
     if(update?.affectedRows > 0) {
